Modularização do código, separando dados e funções em arquivos distintos.
