+++ Chapolin-JS/dados/frasesClassicas.js
@@ -1,0 +1,9 @@
+module.exports = [
+    'Não contavam com minha astúcia!',
+    'Sigam-me os bons!',
+    'Minhas anteninhas de vinil estão detectando a presença do inimigo',
+    'Palma, palma, não priemos cânico!',
+    'Suspeitei desde o princípio!',
+    'Aproveitam-se de minha nobreza',
+    'Todos os meus movimentos são friamente calculados'
+];
