+++ Chapolin-JS/dados/frasesFinais.js
@@ -1,0 +1,3 @@
+module.exports = [
+    'Até a próxima, neste mesmo canal e no mesmo horário!'
+];
