+++ Chapolin-JS/dados/objetos.js
@@ -1,0 +1,6 @@
+module.exports = [
+    'Marreta Biônica',
+    'Aerolito',
+    'Caderninho de anotações',
+    'Chirrin Chirrion do Diabo'
+];
