+++ Chapolin-JS/dados/viloes.js
@@ -1,0 +1,8 @@
+module.exports = [
+    'Poucas Trancas',
+    'Quase Nada',
+    'Tripa Seca',
+    'Racha Cuca',
+    'Pirata Alma Negra',
+    'Rosa Rumorosa'
+];
